Refactor to reorg table actions
