+++ src/lib/utilities/DataTable/DataTable.ts
@@ -4,7 +4,10 @@
 import { type Writable, get } from 'svelte/store';
 import type { DataTableModel } from './types';
 
+// Exports
+
 export * from './types';
+export * from './actions';
 
 // Data Table Handler
 
@@ -105,97 +108,4 @@ function sortOrder(order: string, store: DataTableModel): void {
 			return (x[key] as number) - (y[key] as number);
 		}
 	});
-}
-
-// Actions ---
-
-/** Svelte Action for applying sort asc/dsc classes. */
-export function dataTableInteraction(node: HTMLElement) {
-	const classAsc: string = 'table-sort-asc';
-	const classDsc: string = 'table-sort-dsc';
-	// Click Handler
-	const onClick = (e: any) => {
-		if (!(e.target instanceof Element)) return;
-		// Clear asc class
-		const elemAsc = node.querySelector(`.${classAsc}`);
-		if (elemAsc) elemAsc.classList.remove(classAsc);
-		// Clear dsc class
-		const elemDsc = node.querySelector(`.${classDsc}`);
-		if (elemDsc) elemDsc.classList.remove(classDsc);
-		// Add asc class to data-sort target
-		if (e.target.getAttribute('data-sort')) {
-			const classToApply = sortState.asc ? classAsc : classDsc;
-			e.target.classList.add(classToApply);
-		}
-	};
-	// Events
-	node.addEventListener('click', onClick);
-	// Lifecycle
-	return {
-		destroy() {
-			node.removeEventListener('click', onClick);
-		}
-	};
-}
-
-/** Svelte Action for handling table a11y keyboard interactions. */
-export function tableA11y(node: HTMLElement) {
-	const keyWhitelist: string[] = ['ArrowRight', 'ArrowUp', 'ArrowLeft', 'ArrowDown', 'Home', 'End'];
-	// on:keydown
-	const onKeyDown = (event: KeyboardEvent) => {
-		// console.log('keydown triggered');
-		if (keyWhitelist.includes(event.code)) {
-			event.preventDefault();
-			// prettier-ignore
-			switch (event.code) {
-				case 'ArrowUp': a11ySetActiveCell(node, 0, -1); break;
-				case 'ArrowDown': a11ySetActiveCell(node, 0, 1); break;
-				case 'ArrowLeft': a11ySetActiveCell(node, -1, 0); break;
-				case 'ArrowRight': a11ySetActiveCell(node, 1, 0); break;
-				case 'Home': a11yJumpToOuterColumn(node, 'first'); break;
-				case 'End': a11yJumpToOuterColumn(node, 'last'); break;
-				default: break;
-			}
-		}
-	};
-	// Event Listner
-	node.addEventListener('keydown', onKeyDown);
-	// Lifecycle
-	return {
-		destroy() {
-			node.removeEventListener('keydown', onKeyDown);
-		}
-	};
-}
-
-function a11ySetActiveCell(node: HTMLElement, x: number, y: number): void {
-	// Focused Element
-	const focusedElem = document.activeElement;
-	if (!focusedElem || !focusedElem.parentElement || !focusedElem.parentElement.ariaRowIndex || !focusedElem.ariaColIndex) return;
-	const focusedElemRowIndex = parseInt(focusedElem.parentElement.ariaRowIndex);
-	const focusedElemColIndex = parseInt(focusedElem.ariaColIndex);
-	// Target Element
-	const targetRowElement = node.querySelector(`[aria-rowindex="${focusedElemRowIndex + y}"]`);
-	if (targetRowElement !== null) {
-		const targetColElement: HTMLElement | null = targetRowElement.querySelector(`[aria-colindex="${focusedElemColIndex + x}"]`);
-		if (targetColElement !== null) targetColElement.focus();
-	}
-}
-function a11yGetTargetElem(node: HTMLElement): HTMLElement | null {
-	// Focused Element
-	const focusedElem = document.activeElement;
-	if (!focusedElem || !focusedElem.parentElement || !focusedElem.parentElement.ariaRowIndex) return null;
-	const focusedElemRowIndex = parseInt(focusedElem.parentElement.ariaRowIndex);
-	// Return Target Element
-	return node.querySelector(`[aria-rowindex="${focusedElemRowIndex}"]`);
-}
-
-function a11yJumpToOuterColumn(node: HTMLElement, type: 'first' | 'last' = 'first'): void {
-	const targetRowElement = a11yGetTargetElem(node);
-	if (targetRowElement === null) return;
-	const lastIndex = targetRowElement.children.length;
-	const selected = type === 'first' ? 1 : lastIndex;
-	const targetColElement: HTMLElement | null = targetRowElement.querySelector(`[aria-colindex="${selected}"]`);
-	if (targetColElement === null) return;
-	targetColElement.focus();
 }
